lavapayments/requirements.txt
# requirements.txt
# Install with: pip install -r requirements.txt

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.40.0

# Google Calendar integration (keep your existing setup)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Optional: for better JSON handling
# orjson>=3.9.0

# Optional: for testing
# pytest>=7.4.0
# pytest-mock>=3.11.0